mcp/requirements.txt
# MCP Backend Dependencies
# =========================
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP client for API calls
aiohttp==3.9.1
httpx==0.25.2

# File handling
aiofiles==23.2.0
PyYAML==6.0.1

# System monitoring (optional)
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Logging and utilities
python-multipart==0.0.6